fixed userid integer bug

+++ src/main/java/edu/sjsu/LPOS/repository/UserRepository.java
@@ -9,5 +9,5 @@ import edu.sjsu.LPOS.domain.User;
 public interface UserRepository extends CrudRepository<User, Integer>{
 	
 	User findByUsername(String username);
-	User findByUserId(long userid);
+	User findByUserId(int userid);
 }
